requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-groq>=0.1.0
pydantic>=2.6.0
python-dotenv>=1.0.0
requests>=2.31.0
rich>=13.7.0
typer>=0.12.0
